+++ packages/rum-core/src/common/compress.js
@@ -173,7 +173,9 @@ export function compressTransaction(transaction) {
       t: span.type,
       s: span.start,
       d: span.duration,
-      c: compressContext(span.context)
+      c: compressContext(span.context),
+      o: span.outcome,
+      sr: span.sample_rate
     }
     /**
      * Set parentId only for spans that are child of other spans
@@ -210,7 +212,9 @@ export function compressTransaction(transaction) {
     yc: {
       sd: spans.length
     },
-    sm: transaction.sampled
+    sm: transaction.sampled,
+    sr: transaction.sample_rate,
+    o: transaction.outcome
   }
 
   if (transaction.experience) {