+++ packages/rum-core/src/performance-monitoring/span-base.js
@@ -49,10 +49,12 @@ class SpanBase {
     this.id = options.id || generateRandomId(16)
     this.traceId = options.traceId
     this.sampled = options.sampled
+    this.sampleRate = options.sampleRate
     this.timestamp = options.timestamp
     this._start = getTime(options.startTime)
     this._end = undefined
     this.ended = false
+    this.outcome = undefined
     this.onEnd = options.onEnd
   }