+++ packages/rum-core/test/common/apm-server.spec.js
@@ -131,30 +131,6 @@ describe('ApmServer', function() {
     clock.uninstall()
   })
 
-  it('should capture errors logs from apm-server', done => {
-    apmServer.init()
-    spyOn(loggingService, 'warn').and.callFake((failedMsg, error) => {
-      expect(failedMsg).toEqual('Failed sending events!')
-      /**
-       * APM server error varies by stack, So we check for
-       * explicit characters instead of whole message
-       */
-      expect(error.message).toContain(
-        ': I[#] S[#] doesn\'t validate with "transaction#'
-      )
-      expect(error.message).toContain('missing properties: "trace_id"')
-      done()
-    })
-
-    apmServer.addTransaction({
-      id: '21312',
-      span_count: 0,
-      duration: 100,
-      type: 'app'
-    })
-    apmServer.queue.flush()
-  })
-
   it('should log parse error when response is invalid', done => {
     spyOn(loggingService, 'debug').and.callFake(message => {
       expect(message).toEqual('Error parsing response from APM server')
@@ -360,7 +336,7 @@ describe('ApmServer', function() {
     const expected = [
       '{"m":{"se":{"n":"test","a":{"n":"rum-js","ve":"N/A"},"la":{"n":"javascript"}}}}',
       '{"e":{"id":"error-id-0","cl":"(inline script)","ex":{"mg":"error #0","st":[]},"c":null}}',
-      '{"x":{"id":"transaction-id-0","tid":"trace-id-0","n":"transaction #0","t":"transaction","d":990,"c":null,"m":null,"me":[{"sa":{"xdc":{"v":1},"xds":{"v":990},"xbc":{"v":1}}},{"y":{"t":"app"},"sa":{"ysc":{"v":1},"yss":{"v":980}}},{"y":{"t":"type"},"sa":{"ysc":{"v":1},"yss":{"v":10}}}],"y":[{"id":"span-id-0-1","n":"name","t":"type","s":10,"d":10,"c":null,"su":"subtype"}],"yc":{"sd":1},"sm":true}}'
+      '{"x":{"id":"transaction-id-0","tid":"trace-id-0","n":"transaction #0","t":"transaction","d":990,"c":null,"m":null,"me":[{"sa":{"xdc":{"v":1},"xds":{"v":990},"xbc":{"v":1}}},{"y":{"t":"app"},"sa":{"ysc":{"v":1},"yss":{"v":980}}},{"y":{"t":"type"},"sa":{"ysc":{"v":1},"yss":{"v":10}}}],"y":[{"id":"span-id-0-1","n":"name","t":"type","s":10,"d":10,"c":null,"sr":0.1,"su":"subtype"}],"yc":{"sd":1},"sm":true,"sr":0.1}}'
     ]
     expect(payload.split('\n').filter(a => a)).toEqual(expected)
     clock.uninstall()
@@ -389,8 +365,8 @@ describe('ApmServer', function() {
     const result = apmServer.ndjsonTransactions(payload)
     /* eslint-disable max-len */
     const expected = [
-      '{"transaction":{"id":"transaction-id-0","trace_id":"trace-id-0","name":"transaction #0","type":"transaction","duration":990,"span_count":{"started":1},"sampled":true}}\n',
-      '{"span":{"id":"span-id-0-1","transaction_id":"transaction-id-0","parent_id":"transaction-id-0","trace_id":"trace-id-0","name":"name","type":"type","subtype":"subtype","sync":false,"start":10,"duration":10}}\n',
+      '{"transaction":{"id":"transaction-id-0","trace_id":"trace-id-0","name":"transaction #0","type":"transaction","duration":990,"span_count":{"started":1},"sampled":true,"sample_rate":0.1}}\n',
+      '{"span":{"id":"span-id-0-1","transaction_id":"transaction-id-0","parent_id":"transaction-id-0","trace_id":"trace-id-0","name":"name","type":"type","subtype":"subtype","sync":false,"start":10,"duration":10,"sample_rate":0.1}}\n',
       '{"metricset":{"transaction":{"name":"transaction #0","type":"transaction"},"samples":{"transaction.duration.count":{"value":1},"transaction.duration.sum.us":{"value":990},"transaction.breakdown.count":{"value":1}}}}\n',
       '{"metricset":{"transaction":{"name":"transaction #0","type":"transaction"},"span":{"type":"app"},"samples":{"span.self_time.count":{"value":1},"span.self_time.sum.us":{"value":980}}}}\n',
       '{"metricset":{"transaction":{"name":"transaction #0","type":"transaction"},"span":{"type":"type","subtype":"subtype"},"samples":{"span.self_time.count":{"value":1},"span.self_time.sum.us":{"value":10}}}}\n'
@@ -549,5 +525,62 @@ describe('ApmServer', function() {
     },
     testConfig.stackVersion &&
       compareVersions(testConfig.stackVersion, '7.8.0') >= 0
+  )
+
+  describeIf(
+    '7.10-',
+    () => {
+      it('should capture errors logs from apm-server', done => {
+        apmServer.init()
+        spyOn(loggingService, 'warn').and.callFake((failedMsg, error) => {
+          expect(failedMsg).toEqual('Failed sending events!')
+          /**
+           * APM server error varies by stack, So we check for
+           * explicit characters instead of whole message
+           */
+          expect(error.message).toContain(
+            ': I[#] S[#] doesn\'t validate with "transaction#'
+          )
+          expect(error.message).toContain('missing properties: "trace_id"')
+          done()
+        })
+
+        apmServer.addTransaction({
+          id: '21312',
+          span_count: 0,
+          duration: 100,
+          type: 'app'
+        })
+        apmServer.queue.flush()
+      })
+    },
+    testConfig.stackVersion &&
+      compareVersions(testConfig.stackVersion, '7.10.0') < 0
+  )
+
+  describeIf(
+    '7.10+',
+    () => {
+      it('should capture errors logs from apm-server', done => {
+        apmServer.init()
+        spyOn(loggingService, 'warn').and.callFake((failedMsg, error) => {
+          expect(failedMsg).toEqual('Failed sending events!')
+          expect(error.message).toContain(
+            `validation error: transaction: 'trace_id' required`
+          )
+          done()
+        })
+
+        apmServer.addTransaction({
+          id: '21312',
+          span_count: { started: 0 },
+          duration: 100,
+          type: 'app'
+        })
+        apmServer.queue.flush()
+      })
+    },
+    testConfig.stackVersion &&
+      compareVersions(testConfig.stackVersion, '7.10.0') >= 0
   )
 })